Add files via upload
Added  shifters to find shifted signExtend  and to extend instructions[25-0]. An adder was use to find the branch address through adding the shifted signExtend and PC+4. One mux to decide whether to outpute branch or PC+4 to final mux with s_beq as a signal from the control unit. Another mux to decide whether to use rs for jr or the jump address from s_SignShift for the final mux. Final mux decide wheter to take a (branch or PC+4) or (j or jr) base on s_Jump from the control unit.